cast to int

+++ aclint_checker/aclint_checker.cpp
@@ -357,8 +357,9 @@ extern "C" int get_hart_enable_ids_from_plusargs(int* result, const char* plusar
         numbers.push_back(t);
       }
     }
+
     for (size_t i = 0; i < numbers.size(); ++i) {
-        if (i < NHARTS) {
+        if ((int)i < NHARTS) {
             result[i] = numbers[i];
         }
     }
